Huge Update!!!
Got the GUI up and Running!!!!!
Has a demo of DFS that can be run on it, now working on Dijkstra's.
Has a fully animated graph that can be stepped through one traversal at a time. Included a photo of what the graph looks like, still working on a campus model, so a campus map is acting as the placeholder